% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{block_degree_sequence}
\alias{block_degree_sequence}
\title{Sequence of effective degrees per block.}
\usage{
block_degree_sequence(graph, partition, n.blocks = NULL)
}
\arguments{
\item{graph}{An igraph graph.}

\item{partition}{A vector of group id's for each node of graph.}

\item{n.blocks}{The number of groups. Inferred from the
number of individual group id's in partition by default.}
}
\description{
Sequence of effective degrees per block.
}
