% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{block_edge_counts}
\alias{block_edge_counts}
\title{Matrix of edge counts between blocks}
\usage{
block_edge_counts(graph, partition)
}
\arguments{
\item{graph}{An igraph graph.}

\item{partition}{A vector of group id's for each node of graph.}
}
\description{
Matrix of edge counts between blocks
}
