% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{calculate_dl}
\alias{calculate_dl}
\title{Description length of a block partition}
\usage{
calculate_dl(graph, partition, degree_correction = FALSE)
}
\arguments{
\item{graph}{An igraph graph.}

\item{partition}{Vector of integer values giving the block membership of each
vertex}

\item{degree_correction}{Whether to use degree correction in the calculation.}
}
\value{
Description length value (numeric) for the given graph and partition.
}
\description{
Calculate the description length associated with the current block partition
and type of degree correction.
}
\examples{
g <- sample_ppm(10, 0.9, 10, 3)
calculate_dl(g)
}
\keyword{block}
\keyword{correction}
\keyword{degree}
\keyword{graphs,}
\keyword{inference,}
\keyword{model,}
\keyword{stochastic}
