% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_partition}
\alias{check_partition}
\title{Check partition}
\usage{
check_partition(partition)
}
\arguments{
\item{partition}{A vector of group id's for each node of graph.}
}
\value{
A vector of group id's going from 1 to B.
}
\description{
Renumber partitions from 1 to the number of unique blocks (B).
}
