% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_step}
\alias{collapse_step}
\title{Perform an agglomerative merge step}
\usage{
collapse_step(
  graph,
  partition,
  degree_correction = FALSE,
  n.merges = 1,
  n.moves = 10,
  n.sweeps = 0,
  eps = 0.1,
  beta = 1,
  verbose = TRUE
)
}
\arguments{
\item{graph}{An igraph graph.}

\item{partition}{An vector of integers giving the block partition of nodes.}

\item{degree_correction}{Whether to use degree correction.}

\item{n.merges}{Number of block merges in collapse step.}

\item{n.moves}{Number of merge trials per block.}

\item{n.sweeps}{Number of sweeps after block merge.}

\item{eps}{A parameter controlling ...}

\item{beta}{A parameter controlling ....}

\item{verbose}{Wether to print verbose output to console.}
}
\description{
Collapse \code{n.merges} blocks together via agglomerative merging. For each
block, \code{n.moves} merges are proposed using the same move proposal algorithm
used for the MCMC sweeps. The move proposals are ranked and \code{n.merges}
moves are performed after accounting for duplicates. After the merge, \code{n.sweeps}
MCMC sweeps are performed to equilibrate the resulting partition.
}
