% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsbm.R, R/model.R
\docType{package}
\name{dcsbm}
\alias{dcsbm}
\title{dcsbm: Estimate degree corrected stochastic block models}
\usage{
dcsbm(
  graph,
  degree_correction = FALSE,
  n.blocks = c(1, Inf),
  n.moves = 10,
  n.sweeps = 0,
  verbose = TRUE,
  control = list(sigma = 1.5, eps = 0.1, beta = 1, start_partition = NULL)
)
}
\arguments{
\item{graph}{An igraph graph.}

\item{degree_correction}{Whether to use degree correction.}

\item{n.blocks}{Number of blocks.}

\item{n.moves}{Number of merge trials per block.}

\item{n.sweeps}{Number of sweeps after block merge.}

\item{verbose}{Wether to print verbose output during estimation.}

\item{control}{List of parameters for the inference algorithm.}
}
\value{
Todo
}
\description{
The 'dcsbm' packages provides methods for estimating a two-way degree
corrected stochastic block model for directed, weighted graphs. Depends on
the 'igraph' library \url{https://igraph.org} for graph handling. See Peixoto
(2014) <10.1103/PhysRevE.89.012804> for details on the inference algorithm.

Estimate a stochastic block model with or without degree correction. Works
for directed multigraphs with self-loops.
}
\details{
Estimate a degree corrected stochastic block model: \code{\link{dcsbm}}

Sample a (weighted) graph from the planted partition model: \code{\link{sample_ppm}}
Sample a (weighted) graph from a degree variable planted partition model: \code{\link{sample_dcppm}}
}
\examples{
## Three groups with weighted connections.
g1 <- sample_ppm(60, 0.9, 10, 3)
model <- dcsbm(g1)
}
\keyword{block}
\keyword{correction}
\keyword{degree}
\keyword{graphs,}
\keyword{inference,}
\keyword{model,}
\keyword{stochastic}
