% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{edge_probs}
\alias{edge_probs}
\title{The probability of ...}
\usage{
edge_probs(E)
}
\arguments{
\item{E}{Block adjacency matrix}
}
\value{
A B x B matrix with rows giving the edge probabilities per group.
}
\description{
The probability of ...
}
