% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy_corrected_oneway}
\alias{entropy_corrected_oneway}
\title{Entropy (degree corrected)}
\usage{
entropy_corrected_oneway(E, d, directed = FALSE)
}
\arguments{
\item{E}{Integer matrix of edge counts associated with current partition.}

\item{d}{Integer vector of node degrees.}

\item{directed}{Whether graph is directed.}
}
\description{
Calculate the entropy associated with the current block partition.
}
\details{
\deqn{
  S_c = - E - \sum_k N_k \ln k! - \frac{1}{2} \sum_{rs} e_{rs} \ln\left(\frac{e_{rs}}{e_r e_s}\right)
}{
  St = - E - sum(N[k] * ln k!) - 0.5 * sum(e[r,s] * ln(e[r,s] / (e[r]*e[s])))
}
}
