% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy_trad}
\alias{entropy_trad}
\title{Entropy (no degree correction)}
\usage{
entropy_trad(E, n, directed = FALSE, simple = TRUE)
}
\arguments{
\item{E}{Integer matrix of edge counts associated with current partition}

\item{n}{Integer vector of node counts associated with current partition}

\item{directed}{Whether graph is directed}

\item{simple}{Whether graph is simple.}
}
\description{
Calculate the entropy associated with the current block partition.
}
\details{
\deqn{S_t = \frac{1}{2} \sum_{rs} n_r n_s H_b\left(\frac{e_{rs}}{n_r n_s}\right)}{St = 0.5 * sum_(n[r] * n[s] * H_b(E[r,s] / (n[r]*n[s]))}
}
