% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Entropy of a block partition}
\usage{
get_entropy(graph, partition, degree_correction = FALSE)
}
\arguments{
\item{graph}{An igraph graph.}

\item{partition}{Vector of integer values giving the block membership of each
vertex}

\item{degree_correction}{Whether to account for degree variability.}
}
\value{
Entropy value (numeric) for the given graph and partition.
}
\description{
Calculate the entropy associated with the current block partition and type
of degree correction.
}
\keyword{block}
\keyword{correction}
\keyword{degree}
\keyword{graphs,}
\keyword{inference,}
\keyword{model,}
\keyword{stochastic}
