% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sweep.R
\name{get_proposal_results}
\alias{get_proposal_results}
\title{Calculate entropy delta of the SBM before and after the proposed move and
the ratio of the probabilities of moving to the proposed block before the move and
moving back to the original block after the move.}
\usage{
get_proposal_results(
  curr_v,
  proposed_new_block,
  old_entropy,
  G,
  old_partition,
  block.edges,
  eps = 0.1,
  beta = 1
)
}
\description{
Calculate entropy delta of the SBM before and after the proposed move and
the ratio of the probabilities of moving to the proposed block before the move and
moving back to the original block after the move.
}
