% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{H_binary}
\alias{H_binary}
\title{Binary entropy function}
\usage{
H_binary(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric, vector or matrix with numbers between 0 and 1.}
}
\description{
Binary entropy function
}
