% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_step}
\alias{mcmc_step}
\title{A single MCMC step over one node}
\usage{
mcmc_step(vertex, graph, partition, PE, Pr, B, eps = 1)
}
\arguments{
\item{vertex}{A node in graph.}

\item{graph}{An igraph graph.}

\item{partition}{Vector of integer values giving the block membership of each
vertex}

\item{PE}{asdf}

\item{Pr}{asdf}

\item{B}{Number of groups to sample from}

\item{eps}{(optional) A number giving the ...}
}
\value{
A new group membership for vertex.
}
\description{
A single MCMC step over one node
}
