% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_step3}
\alias{mcmc_step3}
\title{A single MCMC step over one node (alternative 3)}
\usage{
mcmc_step3(vertex, graph, partition, block.edges, Pr, B, eps = 1)
}
\arguments{
\item{vertex}{A node in graph.}

\item{graph}{An igraph graph.}

\item{partition}{Vector of integer values giving the block membership of each
vertex}

\item{block.edges}{asdf}

\item{Pr}{asdf}

\item{B}{Number of groups to sample from}

\item{eps}{(optional) A number giving the ...}
}
\value{
A new group membership for vertex.
}
\description{
A single MCMC step over one node (alternative 3)
}
