% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sweep.R
\name{mcmc_sweep}
\alias{mcmc_sweep}
\title{Run multiple MCMC sweep overs nodes}
\usage{
mcmc_sweep(G, p, B, dc = FALSE, n.sweeps = 1, eps = 0.1, beta = 1)
}
\arguments{
\item{G}{An igraph graph.}

\item{p}{Vector of integer values giving the initial block membership
of each vertex}

\item{B}{Number of blocks.}

\item{dc}{Whether to use degree correction.}

\item{n.sweeps}{Number of sweeps to run.}

\item{eps}{(optional) A number controlling the randomness of moves.}

\item{beta}{(optional) A number controlling the greediness of moves.}
}
\value{
A list object including the partition state after the last move, as well
as the partition state with the lowest entropy.
}
\description{
Runs a n.sweeps MCMC sweep across all nodes (for algorithm details see Piexoto,
2014). Each node is given a chance to move blocks or stay in current block
and all nodes are processed in random order for each sweep.
}
