% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sweep.R
\name{propose_move}
\alias{propose_move}
\title{A single move proposal during an MCMC sweep}
\usage{
propose_move(curr_v, G, p, block.edges, eps = 1)
}
\arguments{
\item{curr_v}{A node in graph.}

\item{G}{An igraph graph.}

\item{p}{Vector of integer values giving the block membership of each
vertex}

\item{block.edges}{asdf}

\item{eps}{(optional) A number giving the ...}
}
\value{
A new group membership for vertex.
}
\description{
A single move proposal during an MCMC sweep
}
