% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{random_probs}
\alias{random_probs}
\title{The probability of sampling a fully random (or new) group in a single MCMC step}
\usage{
random_probs(E, eps = 1)
}
\arguments{
\item{E}{Block adjacency matrix}

\item{eps}{(optional) A number giving the ...}
}
\value{
A vector of length B giving the above probability per group.
}
\description{
The probability of sampling a fully random (or new) group in a single MCMC step
}
