% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample}
\alias{resample}
\title{Safer resample}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{Numeric or vector from which to resample.}

\item{...}{other arguments of \code{sample}.}
}
\description{
Safer resample
}
