% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_at_least_once}
\alias{sample_at_least_once}
\title{Sample at least once with replacement
See https://stackoverflow.com/a/26350427/10042003.}
\usage{
sample_at_least_once(x, size, prob = NULL)
}
\arguments{
\item{x}{a numeric or vector from which to sample.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{prob}{a vector of probability weights for obtaining the elements of the
vector being sampled.}
}
\value{
A vector of length \code{size} with elements drawn from \code{x} where
every element of \code{x} occurs at least once if \code{size} is larger than
\code{length(x)}.
}
\description{
Sample at least once with replacement
See https://stackoverflow.com/a/26350427/10042003.
}
