% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sbm}
\alias{sbm}
\title{Stochastic block model}
\usage{
sbm(
  graph,
  n.blocks = c(1, Inf),
  n.moves = 10,
  n.sweeps = 0,
  control = list(sigma = 1.5, eps = 0.1, beta = 1, start_partition = NULL)
)
}
\arguments{
\item{graph}{An igraph graph.}

\item{n.blocks}{Number of blocks.}

\item{n.moves}{Number of merge trials per block.}

\item{n.sweeps}{Number of sweeps after block merge.}

\item{control}{List of parameters for the inference algorithm.}
}
\value{
Todo
}
\description{
Estimate a ...
}
\examples{
## Three groups with weighted connections.
g1 <- sample_ppm(60, 0.9, 10, 3)
g1 <- simplify(g1)
g1 <- delete.vertices(g1, degree(g1) == 0)
model <- sbm(g1)
}
\keyword{block}
\keyword{correction}
\keyword{degree}
\keyword{graphs,}
\keyword{inference,}
\keyword{model,}
\keyword{stochastic}
