% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sweep.R
\name{swap_blocks}
\alias{swap_blocks}
\title{Swap the block membership of a vertex}
\usage{
swap_blocks(p, curr_v, proposed_new_block)
}
\arguments{
\item{p}{Vector of integer values giving the block membership of each
vertex}

\item{curr_v}{vertex id}

\item{proposed_new_block}{block id}
}
\description{
Swap the block membership of a vertex
}
