% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sweep.R
\name{update_block_edge_list}
\alias{update_block_edge_list}
\title{Create list of edges per block}
\usage{
update_block_edge_list(block.edges, v, G, new.b, old.b)
}
\arguments{
\item{block.edges}{...}

\item{v}{vertex id}

\item{G}{An igraph graph.}

\item{new.b}{...}

\item{old.b}{...}
}
\value{
A new block.edges list.
}
\description{
Create list of edges per block
}
