% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{block_node_counts}
\alias{block_node_counts}
\title{Node counts for blocks}
\usage{
block_node_counts(partition, n.blocks = NULL)
}
\arguments{
\item{partition}{A vector of group id's for each node of graph.}

\item{n.blocks}{The number of groups. Inferred from the
number of individual group id's in partition by default.}
}
\description{
Node counts for blocks
}
