% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsbm.R
\docType{package}
\name{dcsbm}
\alias{dcsbm}
\title{dcsbm: Estimate degree corrected stochastic block models}
\description{
The 'dcsbm' packages provides methods for estimating a two-way degree
corrected stochastic block model for directed, weighted graphs. Depends on
the 'igraph' library \url{https://igraph.org} for graph handling. See Peixoto
(2014) <10.1103/PhysRevE.89.012804> for details on the inference algorithm.
}
\details{
Estimate a degree corrected stochastic block model: \code{\link{dcsbm}}

Sample a (weighted) graph from the planted partition model: \code{\link{sample_ppm}}
}
