% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Entropy of a block partition}
\usage{
get_entropy(graph, partition, degree_correction = FALSE)
}
\arguments{
\item{graph}{An igraph graph.}

\item{partition}{Vector of integer values giving the block membership of each
vertex}

\item{degree_correction}{Whether to account for degree variability.}
}
\value{
Entropy value (numeric) for the given graph and partition.
}
\description{
Calculate the entropy associated with the current block partition and type
of degree correction.
}
\examples{
g <- sample_ppm(30, 0.9, 10, 3)
# Entropy of a random partition
p_random <- sample(1:3, size = 30, replace = TRUE)
get_entropy(g, p_random)
# Entropy of the true partition
p_true <- c(rep(1,30), rep(2,30), rep(3,30))
get_entropy(g, p_true)
}
\keyword{block}
\keyword{correction}
\keyword{degree}
\keyword{graphs,}
\keyword{inference,}
\keyword{model,}
\keyword{stochastic}
