% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Entropy of a block partition}
\usage{
get_entropy(
  graph,
  partition,
  n.blocks = NULL,
  degree_correction = c("none", "oneway", "twoways")
)
}
\arguments{
\item{graph}{An igraph graph.}

\item{partition}{Vector of integer values giving the block membership of each
vertex}

\item{n.blocks}{The number of groups. Inferred from the
number of individual group id's in partition by default.}

\item{degree_correction}{Type of degree correction to use. "oneway" for one
parameter per vertex, "twoways" for two parameters per vertex (input/output),
"none" for no degree correction.}
}
\value{
Entropy value (numeric) for the given graph and partition.
}
\description{
Calculate the entropy associated with the current block partition and type
of degree correction. Can be used with directed and undirected graphs.
}
\examples{
## Three groups with weighted connections.
g1 <- sample_ppm(100, 0.3, 0.03, block.sizes = c(30, 50, 20))
## random partition
p1 <- sample(c(1,2,3), 100, replace = TRUE)
get_entropy(g1, p1)
}
\keyword{block}
\keyword{correction}
\keyword{degree}
\keyword{graphs,}
\keyword{inference,}
\keyword{model,}
\keyword{stochastic}
