% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sweep.R
\name{get_proposal_results}
\alias{get_proposal_results}
\title{Calculate entropy delta of the SBM before and after the proposed move and
the ratio of the probabilities of moving to the proposed block before the move and
moving back to the original block after the move.}
\usage{
get_proposal_results(
  curr_v,
  proposed_new_block,
  old_entropy,
  G,
  old_partition,
  block.edges,
  eps = 0.1,
  beta = 1,
  dc = FALSE
)
}
\arguments{
\item{curr_v}{A node in graph.}

\item{proposed_new_block}{An integer giving the proposed block id.}

\item{old_entropy}{A number giving the pre move entropy}

\item{G}{An igraph graph.}

\item{old_partition}{Vector of integer values giving the block membership of each
vertex pre move.}

\item{block.edges}{a list of all incident edge ids per block}

\item{eps}{(optional) A number giving the ...}

\item{beta}{(optional) A number giving the greediness of the moves.}

\item{dc}{Wether to use degree correction.}
}
\value{
A new group membership for vertex.
}
\description{
Calculate entropy delta of the SBM before and after the proposed move and
the ratio of the probabilities of moving to the proposed block before the move and
moving back to the original block after the move.
}
