% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{H_binary}
\alias{H_binary}
\title{Binary entropy function}
\usage{
H_binary(ers, nr, ns)
}
\arguments{
\item{ers}{edge count between blocks r and s}

\item{nr}{node count of block r}

\item{ns}{node count of block s}
}
\description{
Binary entropy function
}
