% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{h_func}
\alias{h_func}
\title{Mathematicl function used description length calculation}
\usage{
h_func(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The value of \code{h(x)}
}
\description{
Mathematicl function used description length calculation
}
\examples{
h_func(1)
}
