% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_sweep}
\alias{mcmc_sweep}
\title{Run a single MCMC sweep over nodes}
\usage{
mcmc_sweep(graph, partition, B = NULL, eps = 1, beta = 1)
}
\arguments{
\item{graph}{An igraph graph.}

\item{partition}{Vector of integer values giving the block membership of each
vertex}

\item{eps}{(optional) A number giving the ...}
}
\value{
A new partition given as a vector of integer values.
}
\description{
Runs a single MCMC sweep across all nodes (for algorithm details see Piexoto,
2018). Each node is given a chance to move blocks or stay in current block
and all nodes are processed in random order.
}
