% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_ppm}
\alias{sample_ppm}
\title{Sample planted partition model}
\usage{
sample_ppm(partition, p, q)
}
\arguments{
\item{partition}{Vector of group memberships.}

\item{p}{Edge probability between nodes of same group.}

\item{q}{Edge probability between nodes of different groups.}
}
\value{
An igraph graph.
}
\description{
Generate a random graph using the planted partition model
}
\details{
This function samples graphs from a stochastic block model by sampling
each potential transmission with the probabilities given by \code{p} for
in-group transmissions and \code{q} for between group transmissions and
group membership given by the vector \code{partitions}.
}
\examples{

## Three groups with only a few connection between groups
partition <- sample(1:3, 50, replace = TRUE)
g <- sample_ppm(partition, 0.3, 0.01)
g
}
\keyword{graphs,}
\keyword{partition}
\keyword{planted}
\keyword{sample,}
