% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sbm}
\alias{sbm}
\title{Degree corrected stochastic block model}
\usage{
sbm(
  graph,
  n.blocks,
  degree_correction = c("none", "oneway", "twoways"),
  control = list()
)
}
\arguments{
\item{graph}{An igraph graph.}

\item{n.blocks}{Number of blocks.}

\item{degree_correction}{Type of degree correction to use. "oneway" for one
parameter per vertex, "twoways" for two parameters per vertex (input/output),
"none" for no degree correction.}

\item{control}{List of parameters for the inference algorithm.}
}
\value{
Todo
}
\description{
Estimate a ...
}
\examples{
## Three groups with weighted connections.
g1 <- sample_ppm(1000, p=0.1, q=0.003, block.sizes=c(100,600,300))
model <- sbm(g1, 3)
}
\keyword{block}
\keyword{correction}
\keyword{degree}
\keyword{graphs,}
\keyword{inference,}
\keyword{model,}
\keyword{stochastic}
